Fixes broken build of previous patch.
Adds new options to the message connect menu.

- Add all keys of current Message
This will parse the content of the current message (without the part of quoted messages) and open the add downloads dialog to add them to the queue.

- Open Key in Browser
This will open the selected key in the system default browser
For this feature a new configuration option (URI Prefix) can be configured in the Frost miscellaneous option section.  Only when a prefix is configured this menu item entry will appear. This should prevent users of clicking it accidently, when their system default browser is not secure. The same time this option can be used to configure the IP address and port needed to access the node web interface.